# MagaLu/public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "LuizaLabs Logistics Challenge ORDER-API",
    "version": "1.0.0",
    "description": "API para o desafio de integração logística da LuizaLabs"
  },
  "servers": [
    {
      "url": "http://localhost:6868/api",
      "description": "production server"
    }
  ],
  "paths": {
    "/upload": {
      "post": {
        "summary": "Carregar e processar múltiplos arquivos",
        "description": "Permite o carregamento de múltiplos arquivos, processando-os e retornando os resultados agregados.",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "array",
                    "items": {
                      "type": "file",
                      "format": "binary"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Lista de usuários processados a partir dos arquivos carregados."
          },
          "400": {
            "description": "Nenhum arquivo foi carregado ou o número máximo de arquivos excedido."
          }
        }
      }
    },
    "/memory/orders": {
      "get": {
        "summary": "Obter pedidos da memória",
        "description": "Retorna a lista de pedidos armazenados na memória com base nos filtros fornecidos.",
        "parameters": [
          {
            "in": "query",
            "name": "userId",
            "schema": {
              "type": "number"
            },
            "description": "ID do usuário para filtrar os pedidos."
          },
          {
            "in": "query",
            "name": "orderId",
            "schema": {
              "type": "number"
            },
            "description": "ID do pedido para filtrar os pedidos."
          },
          {
            "in": "query",
            "name": "startDate",
            "schema": {
              "type": "string"
            },
            "description": "Data de início para filtrar os pedidos (formato YYYY-MM-DD)."
          },
          {
            "in": "query",
            "name": "endDate",
            "schema": {
              "type": "string"
            },
            "description": "Data de fim para filtrar os pedidos (formato YYYY-MM-DD)."
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de pedidos da memória filtrada."
          },
          "404": {
            "description": "Nenhum pedido encontrado com os filtros fornecidos."
          }
        }
      }
    },
    "/database/orders": {
      "get": {
        "summary": "Obter pedidos do banco de dados",
        "description": "Retorna a lista de pedidos armazenados no banco de dados com base nos filtros fornecidos.",
        "parameters": [
          {
            "in": "query",
            "name": "userId",
            "schema": {
              "type": "number"
            },
            "description": "ID do usuário para filtrar os pedidos."
          },
          {
            "in": "query",
            "name": "orderId",
            "schema": {
              "type": "number"
            },
            "description": "ID do pedido para filtrar os pedidos."
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de pedidos do banco de dados."
          },
          "400": {
            "description": "Nenhum filtro foi especificado para a busca."
          },
          "404": {
            "description": "Nenhum pedido encontrado com os filtros fornecidos."
          },
          "500": {
            "description": "Erro ao buscar pedidos."
          }
        }
      }
    },
    "/stream-orders": {
      "get": {
        "summary": "Transmitir pedidos em tempo real",
        "description": "Envia os pedidos em tempo real por meio de streaming.",
        "responses": {
          "200": {
            "description": "Pedidos transmitidos com sucesso."
          },
          "400": {
            "description": "Erro ao transmitir pedidos."
          }
        }
      }
    }
  },
  "components": {},
  "tags": []
}